// admin/src/app/gohappygobackend/models/transaction-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TransactionResponseDto {
  amount: number;
  convertedAmount: {
} | null;
  createdAt: string;
  currencyCode: string;
  id: number;
  originalAmount: {
} | null;
  payee: {
} | null;
  payeeId: number;
  payer: {
} | null;
  payerId: number;
  paymentMethod: string;
  request: {
} | null;
  requestId: number;

  /**
   * Whether to show the "Release Funds" button. True if: status is "paid" or "awaiting_transfer", stripeTransferId is null, and user is the payee
   */
  showReleaseFundButton: boolean;
  status: 'pending' | 'paid' | 'awaiting_transfer' | 'awaiting_available_funds' | 'refunded' | 'cancelled';
  stripePaymentIntentId: {
} | null;
  stripeTransferId: {
} | null;

  /**
   * The amount the traveler should receive (before fees and TVA)
   */
  travelerPayment: {
} | null;
  updatedAt: string;
}